+++ mosquito-dto/src/main/java/com/softserve/mosquito/dtos/PriorityDto.java
@@ -4,22 +4,22 @@ package com.softserve.mosquito.dtos;
  * Generic DTO for Priority Entity
  */
 public class PriorityDto {
-    private Byte id;
+    private Long id;
     private String title;
 
     public PriorityDto() {
     }
 
-    public PriorityDto(Byte id, String title) {
+    public PriorityDto(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public Byte getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Byte id) {
+    public void setId(Long id) {
         this.id = id;
     }
