+++ mosquito-entity/src/main/java/com.softserve.mosquito/entities/Status.java
@@ -11,7 +11,7 @@ public final class Status implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Byte id;
+    private Long id;
     private String title;
     @OneToMany(mappedBy = "status")
     private List<Task> tasks = new ArrayList<>();
@@ -19,7 +19,7 @@ public final class Status implements Serializable {
     public Status() {
     }
 
-    public Status(Byte id) {
+    public Status(Long id) {
         this.id = id;
     }
 
@@ -27,16 +27,16 @@ public final class Status implements Serializable {
         this.title = title;
     }
 
-    public Status(Byte id, String title) {
+    public Status(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public Byte getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Byte id) {
+    public void setId(Long id) {
         this.id = id;
     }
